-tx 1、完成一个基本完整的订单流程 2、前台加上订单详情等

+++ web/admin/product/productList.jsp
@@ -180,7 +180,8 @@
                             </td>
                             <td height="20" bgcolor="#FFFFFF">
                                 <div align="center" class="STYLE21">
-                                    <a href="${pageContext.request.contextPath }/admin/productServlet?op=deleteOne&pid=${product.id}">删除</a>
+<%--                                    <a href="${pageContext.request.contextPath }/admin/productServlet?op=deleteOne&pid=${product.id}">删除</a>--%>
+                                    <a href="javascript:if(confirm('确实要删除吗?')) location='${pageContext.request.contextPath }/admin/productServlet?op=deleteOne&pid=${product.id}'">删除</a>
                                     |
                                     <a href="${pageContext.request.contextPath }/admin/productServlet?op=findProductByUpdate&pid=${product.id}">编辑</a>
                                 </div>
